_legacy/requirements.txt
# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-mock==3.12.0
pytest-cov==5.0.0

# Web Framework
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3

# HTTP Client
httpx==0.27.0
aiohttp==3.9.3

# Webhook Security
cryptography==42.0.5

# External APIs
stripe==8.5.0
anthropic==0.21.0

# Database/Convex
requests==2.31.0

# Email & Messaging
aiosmtplib==3.0.1
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3